requirements.txt
Pillow==7.0.0
tesserocr==2.4.0
numpy==1.18.1
opencv-python==4.1.2.30
